refactor: Refactor credential offer controller and service to generate offer with QR code and link
